+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - `<Sparkline />` animation on resize
+- `<LineChart />` first and last labels getting cut off
 
 ## [0.0.12] - 2020-07-27